20240404 크기가 작은 부분 문자열 수정

+++ level1/SmallSubstring.java
@@ -5,14 +5,15 @@ public class SmallSubstring {
     // 이 부분문자열이 나타내는 수가 p가 나타내는 수보다 작거나 같은 것이 나오는 횟수를 return하는 함수 solution을 완성하세요.
     public int solution(String t, String p) {
         int answer = 0;
-        int pNumber = Integer.parseInt(p);
+        long pNumber = Long.parseLong(p);
 
         for (int i = 0; i < t.length() - p.length() + 1; i++) {
-            int subTNumber = Integer.parseInt(t.substring(i, i + p.length()));
+            long subTNumber = Long.parseLong(t.substring(i, i + p.length()));
             if (subTNumber <= pNumber) {
                 answer ++;
             }
         }
         return answer;
     }
+    // 몇몇 케이스에서 런타임 에러가 발생했는데, 큰 숫자들이 있었기 때문에 int가 아닌 long을 사용하니 해결되었다.
 }
